+++ tarantool/src/lib.rs
@@ -69,6 +69,8 @@ pub mod msgpack;
 pub mod net_box;
 pub mod network;
 pub mod proc;
+#[cfg(feature = "picodata")]
+pub mod read_view;
 pub mod schema;
 pub mod sequence;
 pub mod session;